There were 2 secrets to getting this to work: 1. First, we had to declare the declaration of do_rootfs as: fakeroot do_rootfs
This makes all the files entered into the ext4 system
owned by root. This is necessary for a successful boot
of the system.

2. I got complaints about the dpkg's being installed with
the wrong version. I changed PATH to start with /usr/bin so
that we would use the Ubuntu 13.04 dpkg. However, I've seen
this occur a couple of times, and it may just come from running
multistrap more than once. This will require some investigation.