Adjusted android-tools-native to use patches rather than files.
First, we got rid of all the makefiles, etc. we were copying into
the android_tools directory. Instead, we went for a set of patches that
show the differences between what was there and what we need.

Second, we found that android_tools is an illegal name for a package. You
cannot have _ as a character, so we changed it to -.