Fixed MAC Addr handling to decouple eth0 & wlan0
Added if-pre-up.d/eth0-config-macaddr that configures the eth0 MAC addr. It gets installed
with a placeholder, that gets substituted with the real MAC address by the script that pulls
them from Android.
With this change, the script to configure ether MACs, that was being generated by the above
script is no longer needed. That script was also being run every time, which resulted in eth0
being brought down if wlan0 was brought up.

Change-Id: Id82229445217b1cff5a33adaa99ee27b40dce890