Fix the initrd addresses
The initrd wasn't being located correctly since the ramdisk_offset argument to
mkbootimg was being passed earlier on the command line than the base argument. This
causes the ramdisk_offset to be overwritten, since the order of args matters. Fixed
by moving the base argument to the beginning of the arglist
.

Change-Id: I579a7daa8a5d27503178c8a7d5950ad1e505ced2