ATL-3858: Add debiff script for creating deb files patch
Debiff is a tool for getting the patch between two different versions
of the same package. This diff can then be applied to the old version
to generate the new version of the debian package.

Format of the debiff patch:
Header: DEBIFF10 (8 bytes)
Body: Size of control.tar patch 8 bytes, contents of control.tar patch
Size of data.tar patch 8 bytes, contents of data.tar patch.

Use bspatch and dpkg-deb --tar-build to re-assemble the destination deb.

Change-Id: Ibf4be389fb78e5cfb8c926ec9f97d1c9ab775de9