ATL-3858: dpkg: Add new options and support reproducibility of debs
Add options to dpkg for debiff implementation. Debiff is a diff
of two debian packages that can be used later to construct a deb
file based on the old version and the diff from the new version.

Also force the timestamps throughout the deb file, owner and group
IDs to be 0 so that builds become reproducible.

Change-Id: Ic2b0bd3db885c16e632e4b701be96a0e7086de1f