ATL-3858: recovery-image: Add dependencies for dpkg-deb
Add xz, dpkg-deb, and dependency libraries for incremental update
debiff implementation. Dpk-deb is used to build the new .deb file
after the control/data files have been patched. xz is used by
dpkg-deb to create the new .deb file. All the libraries are
dependencies for xz or dpk-deb.

Also adds reboot2fastboot so that the reboot functionality can be
used to boot from the recovery image. This is needed in the two
step upgrade process.

Change-Id: I37914bad5a30f3d52a2c44965d4a8c5187a77b16